+++ onmt/Seq2Seq.lua
@@ -2,28 +2,10 @@
 local Seq2Seq, parent = torch.class('Seq2Seq', 'Model')
 
 local options = {
-  {'-layers', 2,           [[Number of layers in the RNN encoder/decoder]],
-                     {valid=onmt.utils.ExtendedCmdLine.isUInt()}},
-  {'-rnn_size', 500, [[Size of RNN hidden states]],
-                     {valid=onmt.utils.ExtendedCmdLine.isUInt()}},
-  {'-rnn_type', 'LSTM', [[Type of RNN cell]],
-                     {enum={'LSTM','GRU'}}},
   {'-word_vec_size', 0, [[Common word embedding size. If set, this overrides -src_word_vec_size and -tgt_word_vec_size.]],
                      {valid=onmt.utils.ExtendedCmdLine.isUInt()}},
   {'-src_word_vec_size', '500', [[Comma-separated list of source embedding sizes: word[,feat1,feat2,...].]]},
   {'-tgt_word_vec_size', '500', [[Comma-separated list of target embedding sizes: word[,feat1,feat2,...].]]},
-  {'-feat_merge', 'concat', [[Merge action for the features embeddings]],
-                     {enum={'concat','sum'}}},
-  {'-feat_vec_exponent', 0.7, [[When features embedding sizes are not set and using -feat_merge concat, their dimension
-                                will be set to N^exponent where N is the number of values the feature takes.]]},
-  {'-feat_vec_size', 20, [[When features embedding sizes are not set and using -feat_merge sum, this is the common embedding size of the features]],
-                     {valid=onmt.utils.ExtendedCmdLine.isUInt()}},
-  {'-input_feed', 1, [[Feed the context vector at each time step as additional input (via concatenation with the word embeddings) to the decoder.]],
-                     {enum={0,1}}},
-  {'-residual', false, [[Add residual connections between RNN layers.]]},
-  {'-brnn', false, [[Use a bidirectional encoder]]},
-  {'-brnn_merge', 'sum', [[Merge action for the bidirectional hidden states]],
-                     {enum={'concat','sum'}}},
   {'-pre_word_vecs_enc', '', [[If a valid path is specified, then this will load
                                      pretrained word embeddings on the encoder side.
                                      See README for specific formatting instructions.]],
@@ -34,16 +16,26 @@ local options = {
                          {valid=onmt.utils.ExtendedCmdLine.fileNullOrExists}},
   {'-fix_word_vecs_enc', false, [[Fix word embeddings on the encoder side]]},
   {'-fix_word_vecs_dec', false, [[Fix word embeddings on the decoder side]]},
-  {'-dropout', 0.3, [[Dropout probability. Dropout is applied between vertical LSTM stacks.]]}
+  {'-feat_merge', 'concat', [[Merge action for the features embeddings]],
+                     {enum={'concat','sum'}}},
+  {'-feat_vec_exponent', 0.7, [[When features embedding sizes are not set and using -feat_merge concat, their dimension
+                                will be set to N^exponent where N is the number of values the feature takes.]]},
+  {'-feat_vec_size', 20, [[When features embedding sizes are not set and using -feat_merge sum, this is the common embedding size of the features]],
+                     {valid=onmt.utils.ExtendedCmdLine.isUInt()}},
+  {'-input_feed', 1, [[Feed the context vector at each time step as additional input (via concatenation with the word embeddings) to the decoder.]],
+                     {enum={0,1}}}
 }
 
 function Seq2Seq.declareOpts(cmd)
   cmd:setCmdLineOptions(options, Seq2Seq.modelName())
+  onmt.Encoder.declareOpts(cmd)
+  onmt.Factory.declareOpts(cmd)
 end
 
 function Seq2Seq:__init(args, dicts, verbose)
   parent.__init(self, args)
   onmt.utils.Table.merge(self.args, onmt.utils.ExtendedCmdLine.getModuleOpts(args, options))
+  self.args.uneven_batches = args.uneven_batches
 
   self.models.encoder = onmt.Factory.buildWordEncoder(args, dicts.src, verbose)
   self.models.decoder = onmt.Factory.buildWordDecoder(args, dicts.tgt, verbose)
@@ -55,9 +47,10 @@ function Seq2Seq.load(args, models, dicts, isReplica)
 
   parent.__init(self, args)
   onmt.utils.Table.merge(self.args, onmt.utils.ExtendedCmdLine.getModuleOpts(args, options))
+  self.args.uneven_batches = args.uneven_batches
 
-  self.models.encoder = onmt.Factory.loadEncoder(models.encoder, isReplica)
-  self.models.decoder = onmt.Factory.loadDecoder(models.decoder, isReplica)
+  self.models.encoder = onmt.Factory.loadEncoder(models.encoder, isReplica, args)
+  self.models.decoder = onmt.Factory.loadDecoder(models.decoder, isReplica, args)
   self.criterion = onmt.ParallelClassNLLCriterion(onmt.Factory.getOutputSizes(dicts.tgt))
 
   return self
@@ -94,12 +87,26 @@ function Seq2Seq:getOutput(batch)
   return batch.targetOutput
 end
 
+function Seq2Seq:maskPadding(batch)
+  if self.args.uneven_batches then
+    self.models.encoder:maskPadding()
+    if batch.uneven then
+      self.models.decoder:maskPadding(batch.sourceSize, batch.sourceLength)
+    else
+      self.models.decoder:maskPadding()
+    end
+  end
+end
+
 function Seq2Seq:forwardComputeLoss(batch)
+  self:maskPadding(batch)
   local encoderStates, context = self.models.encoder:forward(batch)
   return self.models.decoder:computeLoss(batch, encoderStates, context, self.criterion)
 end
 
 function Seq2Seq:trainNetwork(batch, dryRun)
+  self:maskPadding(batch)
+
   local encStates, context = self.models.encoder:forward(batch)
 
   local decOutputs = self.models.decoder:forward(batch, encStates, context)