+++ onmt/data/Batch.lua
@@ -4,18 +4,22 @@
 local function getLength(seq, ignore)
   local sizes = torch.IntTensor(#seq):zero()
   local max = 0
-  local sum = 0
+  local uneven = false
 
   for i = 1, #seq do
     local len = seq[i]:size(1)
     if ignore ~= nil then
       len = len - ignore
     end
-    max = math.max(max, len)
-    sum = sum + len
+    if max == 0 or len > max then
+      if max ~= 0 then
+        uneven = true
+      end
+      max = len
+    end
     sizes[i] = len
   end
-  return max, sizes, sum
+  return max, sizes, uneven
 end
 
 --[[ Data management and batch creation.
@@ -67,7 +71,7 @@ function Batch:__init(src, srcFeatures, tgt, tgtFeatures)
 
   self.size = #src
 
-  self.sourceLength, self.sourceSize = getLength(src)
+  self.sourceLength, self.sourceSize, self.uneven = getLength(src)
 
   local sourceSeq = torch.LongTensor(self.sourceLength, self.size):fill(onmt.Constants.PAD)
   self.sourceInput = sourceSeq:clone()
@@ -75,6 +79,8 @@ function Batch:__init(src, srcFeatures, tgt, tgtFeatures)
 
   self.sourceInputFeatures = {}
   self.sourceInputRevFeatures = {}
+  -- will be used to return extra padded value
+  self.padTensor = torch.LongTensor(self.size):fill(onmt.Constants.PAD)
 
   if #srcFeatures > 0 then
     for _ = 1, #srcFeatures[1] do
@@ -196,7 +202,13 @@ end
 local function addInputFeatures(inputs, featuresSeq, t)
   local features = {}
   for j = 1, #featuresSeq do
-    table.insert(features, featuresSeq[j][t])
+    local feat
+    if t > featuresSeq[j]:size(1) then
+      feat = onmt.Constants.PAD
+    else
+      feat = featuresSeq[j][t]
+    end
+    table.insert(features, feat)
   end
   if #features > 1 then
     table.insert(inputs, features)
@@ -207,8 +219,14 @@ end
 
 --[[ Get source input batch at timestep `t`. --]]
 function Batch:getSourceInput(t)
+  local inputs
+
   -- If a regular input, return word id, otherwise a table with features.
-  local inputs = self.sourceInput[t]
+  if t > self.sourceInput:size(1) then
+    inputs = self.padTensor
+  else
+    inputs = self.sourceInput[t]
+  end
 
   if #self.sourceInputFeatures > 0 then
     inputs = { inputs }