+++ onmt/modules/Decoder.lua
@@ -168,15 +168,8 @@ end
   * See  [onmt.MaskedSoftmax](onmt+modules+MaskedSoftmax).
 --]]
 function Decoder:maskPadding(sourceSizes, sourceLength)
-  if not self.decoderAttn then
-    self.network:apply(function (layer)
-      if layer.name == 'decoderAttn' then
-        self.decoderAttn = layer
-      end
-    end)
-  end
 
-  self.decoderAttn:replace(function(module)
+  local function substituteSoftmax(module)
     if module.name == 'softmaxAttn' then
       local mod
       if sourceSizes ~= nil then
@@ -192,7 +185,30 @@ function Decoder:maskPadding(sourceSizes, sourceLength)
     else
       return module
     end
-  end)
+  end
+
+  if not self.decoderAttn then
+    self.network:apply(function (layer)
+      if layer.name == 'decoderAttn' then
+        self.decoderAttn = layer
+      end
+    end)
+  end
+  self.decoderAttn:replace(substituteSoftmax)
+
+  if not self.decoderAttnClones then
+    self.decoderAttnClones = {}
+  end
+  for t = 1, #self.networkClones do
+    if not self.decoderAttnClones[t] then
+      self:net(t):apply(function (layer)
+        if layer.name == 'decoderAttn' then
+          self.decoderAttnClones[t] = layer
+        end
+      end)
+    end
+    self.decoderAttnClones[t]:replace(substituteSoftmax)
+  end
 end
 
 --[[ Run one step of the decoder.
@@ -330,7 +346,7 @@ function Decoder:backward(batch, outputs, criterion)
   local gradStatesInput = onmt.utils.Tensor.reuseTensorTable(self.gradOutputsProto,
                                                              { batch.size, self.args.rnnSize })
   local gradContextInput = onmt.utils.Tensor.reuseTensor(self.gradContextProto,
-                                                         { batch.size, batch.sourceLength, self.args.rnnSize })
+                                                         { batch.size, batch.encoderOutputLength or batch.sourceLength, self.args.rnnSize })
 
   local loss = 0
   local indvAvgLoss = torch.zeros(outputs[1]:size(1))