+++ onmt/modules/LSTM.lua
@@ -30,21 +30,22 @@ Parameters:
   * `layers` - Number of LSTM layers, L.
   * `inputSize` - Size of input layer
   * `hiddenSize` - Size of the hidden layers.
-  * `dropout` - Dropout rate to use.
+  * `dropout` - Dropout rate to use (in $$[0,1]$$ range).
   * `residual` - Residual connections between layers.
+  * `dropout_input` - if true, add a dropout layer on the first layer (useful for instance in complex encoders)
 --]]
-function LSTM:__init(layers, inputSize, hiddenSize, dropout, residual)
+function LSTM:__init(layers, inputSize, hiddenSize, dropout, residual, dropout_input)
   dropout = dropout or 0
 
   self.dropout = dropout
   self.numEffectiveLayers = 2 * layers
   self.outputSize = hiddenSize
 
-  parent.__init(self, self:_buildModel(layers, inputSize, hiddenSize, dropout, residual))
+  parent.__init(self, self:_buildModel(layers, inputSize, hiddenSize, dropout, residual, dropout_input))
 end
 
 --[[ Stack the LSTM units. ]]
-function LSTM:_buildModel(layers, inputSize, hiddenSize, dropout, residual)
+function LSTM:_buildModel(layers, inputSize, hiddenSize, dropout, residual, dropout_input)
   local inputs = {}
   local outputs = {}
 
@@ -74,9 +75,9 @@ function LSTM:_buildModel(layers, inputSize, hiddenSize, dropout, residual)
       if residual and (L > 2 or inputSize == hiddenSize) then
         input = nn.CAddTable()({input, prevInput})
       end
-      if dropout > 0 then
-        input = nn.Dropout(dropout)(input)
-      end
+    end
+    if dropout_input or (dropout > 0 and L>1) then
+      input = nn.Dropout(dropout)(input)
     end
 
     local prevC = inputs[L*2 - 1]