+++ onmt/utils/Cuda.lua
@@ -9,7 +9,7 @@ local Cuda = {
 local options = {
   {'-gpuid',     '0',   [[List of comma-separated GPU identifiers (1-indexed). CPU is used when set to 0.]],
                                  {valid=ExtendedCmdLine.listUInt}},
-  {'-fallback_to_cpu', false, [[If GPU can't be use, rollback on the CPU.]]},
+  {'-fallback_to_cpu', false, [[If GPU can't be used, rollback on the CPU.]]},
   {'-fp16', false, [[Use half-precision float on GPU.]]},
   {'-no_nccl', false, [[Disable usage of nccl in parallel mode.]]}
 }