+++ CHANGELOG.md
@@ -3,10 +3,28 @@
 ### New features
 
 * Add TensorBoard metrics visualisation with [Crayon](https://github.com/torrvision/crayon)
+* Add deep bidirectional encoder
+* Add pyramidal deep bidirectional encoder
+* Add alternative learning rate decay strategy for SGD training
+* Add sequence tagger model
+
+### Fixes and improvements
+
+* Fix individual losses that were always computed when using random sampling
+* Allow to change the state of word embeddings optimization during a retraining
+
+## [v0.5.2](https://github.com/OpenNMT/OpenNMT/releases/tag/v0.5.2) (2017-03-29)
+
+### Fixes and improvements
+
+* Improve compatibility with older Torch versions missing the `fmod` implementation
+
+## [v0.5.1](https://github.com/OpenNMT/OpenNMT/releases/tag/v0.5.1) (2017-03-28)
 
 ### Fixes and improvements
 
 * Fix translation with FP16 precision
+* Fix regression that make `tds` mandatory for translation
 
 ## [v0.5.0](https://github.com/OpenNMT/OpenNMT/releases/tag/v0.5.0) (2017-03-06)